backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "anthropic>=0.96.0",
  "faker>=40.15.0",
  "fastapi>=0.136.0",
  "numpy>=2.4.4",
  "pydantic>=2.13.3",
  "pydantic-settings>=2.14.0",
  "python-dateutil>=2.9.0",
  "python-dotenv>=1.2.2",
  "python-jose[cryptography]>=3.5.0",
  "supabase>=2.28.3",
  "uvicorn>=0.45.0",
]

[project.optional-dependencies]
dev = ["pytest", "ruff"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
